backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
httpx==0.27.2
python-multipart==0.0.9
Pillow==11.0.0
opencv-python==4.10.0.84
numpy==2.1.3
# Use versions compatible with Python 3.13 (see pip error log)
torch==2.10.0
torchvision==0.25.0
scikit-learn==1.5.2
albumentations==1.4.8
